# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple todo list API",
        "title": "Todo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/todos": {
            "get": {
                "description": "Get a list of all todo items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Get all todos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_dunstorm_my-svelte-go-template_backend_internal_models.Todo"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new todo item with the provided title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Create a new todo",
                "parameters": [
                    {
                        "description": "Todo object",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_dunstorm_my-svelte-go-template_backend_internal_models.Todo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_dunstorm_my-svelte-go-template_backend_internal_models.Todo"
                        }
                    }
                }
            }
        },
        "/api/todos/{id}": {
            "put": {
                "description": "Update a todo's completion status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Update a todo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Todo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Todo object",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_dunstorm_my-svelte-go-template_backend_internal_models.Todo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dunstorm_my-svelte-go-template_backend_internal_models.Todo"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_dunstorm_my-svelte-go-template_backend_internal_models.Todo": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}